created controller for authentication
